+++ even-more-fish-plugin/src/main/java/com/oheers/fish/recipe/EMFShapedRecipe.java
@@ -1,6 +1,5 @@
 package com.oheers.fish.recipe;
 
-import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.FishUtils;
 import dev.dejvokep.boostedyaml.block.implementation.Section;
 import org.bukkit.NamespacedKey;
@@ -11,12 +10,12 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
-public class EMFShapedRecipe {
+public class EMFShapedRecipe extends EMFRecipe<ShapedRecipe> {
 
-    private final ShapedRecipe recipe;
     private final List<String> rawShape;
     private final Section ingredientsSection;
-    private boolean registered = false;
+    private final NamespacedKey key;
+    private final ItemStack result;
 
     public EMFShapedRecipe(@NotNull NamespacedKey key, @NotNull ItemStack result, @NotNull Section section) {
         this.rawShape = section.getStringList("shape");
@@ -24,11 +23,14 @@ public class EMFShapedRecipe {
         if (this.rawShape.isEmpty() || this.ingredientsSection == null) {
             throw new RuntimeException("Shaped recipe is missing shape or ingredients.");
         }
-        this.recipe = new ShapedRecipe(key, result);
-        prepareRecipe();
+        this.key = key;
+        this.result = result;
     }
 
-    private void prepareRecipe() {
+    @Override
+    protected @NotNull ShapedRecipe prepareRecipe() {
+        ShapedRecipe recipe = new ShapedRecipe(key, result);
+
         String[] shape = rawShape.stream().limit(3).toArray(String[]::new);
         recipe.shape(shape);
 
@@ -48,26 +50,8 @@ public class EMFShapedRecipe {
             }
             recipe.setIngredient(character, choice);
         });
-    }
 
-    public boolean isRegistered() {
-        return this.registered;
-    }
-
-    public void register() {
-        if (isRegistered()) {
-            throw new RuntimeException("Attempted to register a recipe that is already registered.");
-        }
-        EvenMoreFish.getInstance().getServer().addRecipe(recipe);
-        this.registered = true;
-    }
-
-    public void unregister() {
-        if (!isRegistered()) {
-            throw new RuntimeException("Attempted to unregister a recipe that is not registered.");
-        }
-        EvenMoreFish.getInstance().getServer().removeRecipe(recipe.getKey());
-        this.registered = false;
+        return recipe;
     }
 
 }
